[#33] Separated ts client
