+++ envs/.env.alpha
@@ -10,3 +10,7 @@ WS_PORT=3002
 
 SERVER_MODE=alpha
 WEB_MODE=alpha
+
+VITE_PONG_CLIENT_TYPE=http
+VITE_PONG_URL=http://127.0.0.1:3000
+VITE_PONG_WS_URL=http://[::1]:3002
