+++ valtio/.prettierrc
@@ -1,0 +1,8 @@
+{
+  "singleQuote": true,
+  "trailingComma": "all",
+  "semi": true,
+  "printWidth": 80,
+  "jsxSingleQuote": true,
+  "tabWidth": 2
+}
