+++ valtio/src/index.ts
@@ -1,1 +1,2 @@
 export * from './context';
+export * from './react';
