+++ web/package.json
@@ -18,7 +18,8 @@
     "react-router-dom": "6",
     "socket.io-client": "^4.5.1",
     "ts-essentials": "^9.3.0",
-    "valtio": "^1.6.4"
+    "valtio": "^1.6.4",
+    "valtio-pong": "file:./../valtio"
   },
   "devDependencies": {
     "@types/react": "^18.0.17",
